bug fix and refactor
