+++ client/components/index.js
@@ -8,13 +8,9 @@ export {default as MapContainer} from './mapContainer'
 export {Login, Signup} from './authForm'
 export {default as MiniDrawer} from './drawer'
 export {default as NewReservation} from './newReservation'
-export {default as UserAccount} from './userAccount'
-export {default as MyReservations} from './userAccount'
-
+export {default as MyReservations} from './myReservations'
+export {default as TimeLine} from './timeLine'
+export {default as MyProfile} from './myprofile'
+export {default as EditProfile} from './editProfile'
 export {default as LocationScreen} from './locationScreen'
 export {default as EditReservation} from './editReservation'
-
-export {default as Chat} from './chat'
-export {default as EditProfile} from './editProfile'
-
-export {default as TimeLine} from './timeLine'
