--- client/components/myInfo.js
@@ -1,19 +1,0 @@
-import React from 'react'
-import Avatar from '@material-ui/core/Avatar'
-import {Link} from 'react-router-dom'
-import Button from '@material-ui/core/Button'
-
-const MyInfo = props => {
-  return (
-    <div>
-      <Avatar src={props.user.imageUrl} />
-      <div>Name: {props.user.name}</div>
-      <div>Email: {props.user.email}</div>
-      <Link to="/profile">
-        <Button variant="outlined">edit my profile</Button>
-      </Link>
-    </div>
-  )
-}
-
-export default MyInfo
