+++ server/api/messages.js
@@ -13,14 +13,12 @@ router.get('/', async (req, res, next) => {
 
 router.get('/:id', async (req, res, next) => {
   try {
-    console.log(req.params.id)
     const messages = await Message.findAll({
       where: {
         reservationId: req.params.id
       },
       include: [{all: true}]
     })
-    console.log(messages)
     res.json(messages)
   } catch (err) {
     next(err)
